+++ cmd/compute/instance_type/instance_type_show.go
@@ -9,11 +9,11 @@ import (
 	"github.com/spf13/cobra"
 
 	exocmd "github.com/exoscale/cli/cmd"
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
 	"github.com/exoscale/cli/table"
-	exoapi "github.com/exoscale/egoscale/v2/api"
+	"github.com/exoscale/cli/utils"
+	v3 "github.com/exoscale/egoscale/v3"
 )
 
 type instanceTypeShowOutput struct {
@@ -72,29 +72,28 @@ func (c *instanceTypeShowCmd) CmdPreRun(cmd *cobra.Command, args []string) error
 }
 
 func (c *instanceTypeShowCmd) CmdRun(_ *cobra.Command, _ []string) error {
-	ctx := exoapi.WithEndpoint(
-		exocmd.GContext,
-		exoapi.NewReqEndpoint(account.CurrentAccount.Environment, account.CurrentAccount.DefaultZone),
-	)
-
-	t, err := globalstate.EgoscaleClient.FindInstanceType(ctx, account.CurrentAccount.DefaultZone, c.Type)
+	ctx := exocmd.GContext
+	client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(exocmd.DefaultZone))
+	if err != nil {
+		return err
+	}
+	instanceTypes, err := client.ListInstanceTypes(ctx)
 	if err != nil {
 		return err
 	}
 
+	t, err := instanceTypes.FindInstanceTypeByIdOrFamilyAndSize(c.Type)
+	if err != nil {
+		return err
+	}
 	return c.OutputFunc(&instanceTypeShowOutput{
-		ID:     *t.ID,
-		Family: *t.Family,
-		Size:   *t.Size,
-		Memory: *t.Memory,
-		CPUs:   *t.CPUs,
-		GPUs: func() (v int64) {
-			if t.GPUs != nil {
-				v = *t.GPUs
-			}
-			return
-		}(),
-		Authorized: *t.Authorized,
+		ID:         t.ID.String(),
+		Family:     string(t.Family),
+		Size:       string(t.Size),
+		Memory:     t.Memory,
+		CPUs:       t.Cpus,
+		GPUs:       t.Gpus,
+		Authorized: utils.DefaultBool(t.Authorized, false),
 	}, nil)
 }